requirements.txt
dgl-cu102             0.7.0
entrypoints           0.3
filelock              3.0.12
future                0.18.2
googledrivedownloader 0.4
grakel                0.1.8
huggingface-hub       0.0.12
idna                  3.1
ipykernel             6.0.3
ipython               7.25.0
ipython-genutils      0.2.0
ipywidgets            7.6.3
jedi                  0.18.0
Jinja2                3.0.1
jmespath              0.10.0
joblib                1.0.1
jsonschema            3.2.0
jupyter               1.0.0
jupyter-client        6.2.0
jupyter-console       6.4.0
jupyter-core          4.7.1
jupyterlab-pygments   0.1.2
jupyterlab-widgets    1.0.0
langid                1.1.6
littleutils           0.2.2
MarkupSafe            2.0.1
matplotlib-inline     0.1.2
mistune               0.8.4
mkl-fft               1.3.0
mkl-random            1.2.2
mkl-service           2.4.0
nbclient              0.5.3
nbconvert             6.1.0
nbformat              5.1.3
nest-asyncio          1.5.1
networkx              2.5
nose                  1.3.7
notebook              6.4.0
numpy                 1.20.3
ogb                   1.3.1
olefile               0.46
outdated              0.2.1
packaging             21.0
pandas                1.1.4
pandocfilters         1.4.3
parso                 0.8.2
pexpect               4.8.0
pickleshare           0.7.5
Pillow                8.3.1
pip                   21.1.3
prometheus-client     0.11.0
prompt-toolkit        3.0.19
ptyprocess            0.7.0
pycparser             2.20
Pygments              2.9.0
pyOpenSSL             20.0.1
pyparsing             2.4.7
pyrsistent            0.18.0
PySocks               1.7.1
python-dateutil       2.8.2
python-louvain        0.15
pytz                  2021.1
PyYAML                5.4.1
pyzmq                 22.1.0
qtconsole             5.1.1
QtPy                  1.9.0
regex                 2021.8.21
requests              2.26.0
s3transfer            0.5.0
sacremoses            0.0.45
scikit-learn          0.24.2
scipy                 1.6.2
Send2Trash            1.7.1
setuptools            52.0.0.post20210125
six                   1.16.0
terminado             0.10.1
testpath              0.5.0
texttable             1.6.4
threadpoolctl         2.2.0
tokenizers            0.10.3
torch                 1.9.0
torch-cluster         1.5.9
torch-geometric       1.7.2
torch-scatter         2.0.8
torch-sparse          0.6.11
torch-spline-conv     1.2.1
torchaudio            0.9.0a0+33b2469
torchtext             0.10.0
torchvision           0.10.0
tornado               6.1
tqdm                  4.61.2
traitlets             5.0.5
transformers          4.9.2
typing-extensions     3.10.0.0
urllib3               1.26.6
wcwidth               0.2.5
webencodings          0.5.1
wheel                 0.36.2
widgetsnbextension    3.5.1
